Moved bindings into their own source file.
